Refactor second non-C*4 test (one for SAI) into its own test class; create a new intermediate base class
