+++ testing/src/main/java/io/stargate/it/http/restapi/RestApiTestBase.java
@@ -8,18 +8,25 @@ import io.stargate.it.BaseIntegrationTest;
 import io.stargate.it.http.RestUtils;
 import io.stargate.it.http.models.Credentials;
 import io.stargate.it.storage.StargateConnectionInfo;
+import io.stargate.web.restapi.models.ColumnDefinition;
 import io.stargate.web.restapi.models.GetResponseWrapper;
+import io.stargate.web.restapi.models.PrimaryKey;
+import io.stargate.web.restapi.models.TableAdd;
+import io.stargate.web.restapi.models.TableResponse;
 import java.io.IOException;
 import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestInfo;
 
 /** Intermediate IT base class for use by REST API integration tests. */
-public class RestApiTestBase extends BaseIntegrationTest {
+abstract class RestApiTestBase extends BaseIntegrationTest {
   protected String keyspaceName;
   protected String tableName;
   protected static String authToken;
@@ -77,5 +84,65 @@ public class RestApiTestBase extends BaseIntegrationTest {
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
+  }
+
+  protected void createTestTable(
+      String tableName, List<String> columns, List<String> partitionKey, List<String> clusteringKey)
+      throws IOException {
+    TableAdd tableAdd = new TableAdd();
+    tableAdd.setName(tableName);
+
+    List<ColumnDefinition> columnDefinitions =
+        columns.stream()
+            .map(x -> x.split(" "))
+            .map(y -> new ColumnDefinition(y[0], y[1]))
+            .collect(Collectors.toList());
+    tableAdd.setColumnDefinitions(columnDefinitions);
+
+    PrimaryKey primaryKey = new PrimaryKey();
+    primaryKey.setPartitionKey(partitionKey);
+    if (clusteringKey != null) {
+      primaryKey.setClusteringKey(clusteringKey);
+    }
+    tableAdd.setPrimaryKey(primaryKey);
+
+    String body =
+        RestUtils.post(
+            authToken,
+            String.format("%s/v2/schemas/keyspaces/%s/tables", restUrlBase, keyspaceName),
+            objectMapper.writeValueAsString(tableAdd),
+            HttpStatus.SC_CREATED);
+
+    TableResponse tableResponse = objectMapper.readValue(body, TableResponse.class);
+    assertThat(tableResponse.getName()).isEqualTo(tableName);
+  }
+
+  /**
+   * Helper method for inserting table entries using so-called "Stringified" values for columns:
+   * this differs a bit from full JSON values and is mostly useful for simple String and number
+   * fields.
+   *
+   * @return {@code List} of entries to expect back for given definitions.
+   */
+  protected List<Map<String, String>> insertTestTableRows(List<List<String>> rows)
+      throws IOException {
+    final List<Map<String, String>> insertedRows = new ArrayList<>();
+    for (List<String> row : rows) {
+      Map<String, String> rowMap = new HashMap<>();
+      for (String kv : row) {
+        // Split on first space, leave others in (with no limit we'd silently
+        // drop later space-separated parts)
+        String[] parts = kv.split(" ", 2);
+        rowMap.put(parts[0].trim(), parts[1].trim());
+      }
+      insertedRows.add(rowMap);
+
+      RestUtils.post(
+          authToken,
+          String.format("%s/v2/keyspaces/%s/%s", restUrlBase, keyspaceName, tableName),
+          objectMapper.writeValueAsString(rowMap),
+          HttpStatus.SC_CREATED);
+    }
+    return insertedRows;
   }
 }
