+++ testing/src/main/java/io/stargate/it/http/restapi/RestApiv2MVTest.java
@@ -8,19 +8,12 @@ import com.datastax.oss.driver.api.core.cql.ResultSet;
 import io.stargate.it.driver.CqlSessionExtension;
 import io.stargate.it.driver.CqlSessionSpec;
 import io.stargate.it.http.RestUtils;
-import io.stargate.web.restapi.models.ColumnDefinition;
-import io.stargate.web.restapi.models.PrimaryKey;
-import io.stargate.web.restapi.models.TableAdd;
-import io.stargate.web.restapi.models.TableResponse;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 import net.jcip.annotations.NotThreadSafe;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
@@ -34,13 +27,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 @ExtendWith(CqlSessionExtension.class)
 @CqlSessionSpec()
 public class RestApiv2MVTest extends RestApiTestBase {
-
-  /*
-  /************************************************************************
-  /* Test methods
-  /************************************************************************
-   */
-
   @Test
   public void getAllRowsFromMaterializedView(CqlSession session) throws IOException {
     assumeThat(isCassandra4())
@@ -96,71 +82,5 @@ public class RestApiv2MVTest extends RestApiTestBase {
 
     assertThat(rows.size()).isEqualTo(2);
     assertThat(new LinkedHashSet<>(rows)).isEqualTo(new LinkedHashSet<>(expRows));
-  }
-
-  /*
-  /************************************************************************
-  /* Helper methods for setting up tests
-  /************************************************************************
-   */
-
-  private void createTestTable(
-      String tableName, List<String> columns, List<String> partitionKey, List<String> clusteringKey)
-      throws IOException {
-    TableAdd tableAdd = new TableAdd();
-    tableAdd.setName(tableName);
-
-    List<ColumnDefinition> columnDefinitions =
-        columns.stream()
-            .map(x -> x.split(" "))
-            .map(y -> new ColumnDefinition(y[0], y[1]))
-            .collect(Collectors.toList());
-    tableAdd.setColumnDefinitions(columnDefinitions);
-
-    PrimaryKey primaryKey = new PrimaryKey();
-    primaryKey.setPartitionKey(partitionKey);
-    if (clusteringKey != null) {
-      primaryKey.setClusteringKey(clusteringKey);
-    }
-    tableAdd.setPrimaryKey(primaryKey);
-
-    String body =
-        RestUtils.post(
-            authToken,
-            String.format("%s/v2/schemas/keyspaces/%s/tables", restUrlBase, keyspaceName),
-            objectMapper.writeValueAsString(tableAdd),
-            HttpStatus.SC_CREATED);
-
-    TableResponse tableResponse = objectMapper.readValue(body, TableResponse.class);
-    assertThat(tableResponse.getName()).isEqualTo(tableName);
-  }
-
-  /**
-   * Helper method for inserting table entries using so-called "Stringified" values for columns:
-   * this differs a bit from full JSON values and is mostly useful for simple String and number
-   * fields.
-   *
-   * @return {@code List} of entries to expect back for given definitions.
-   */
-  private List<Map<String, String>> insertTestTableRows(List<List<String>> rows)
-      throws IOException {
-    final List<Map<String, String>> insertedRows = new ArrayList<>();
-    for (List<String> row : rows) {
-      Map<String, String> rowMap = new HashMap<>();
-      for (String kv : row) {
-        // Split on first space, leave others in (with no limit we'd silently
-        // drop later space-separated parts)
-        String[] parts = kv.split(" ", 2);
-        rowMap.put(parts[0].trim(), parts[1].trim());
-      }
-      insertedRows.add(rowMap);
-
-      RestUtils.post(
-          authToken,
-          String.format("%s/v2/keyspaces/%s/%s", restUrlBase, keyspaceName, tableName),
-          objectMapper.writeValueAsString(rowMap),
-          HttpStatus.SC_CREATED);
-    }
-    return insertedRows;
   }
 }
